requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
async-timeout==4.0.2
    # via redis
click==8.1.3
    # via flask
colorama==0.4.6
    # via click
flake8==6.0.0
    # via -r requirements-dev.in
flask==2.2.2
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
python-dotenv==0.21.0
    # via -r requirements.in
redis==4.4.0
    # via -r requirements.in
waitress==2.1.2
    # via -r requirements.in
werkzeug==2.2.2
    # via flask
